Refactor JsonCollection class to have inner classes for Document
