+++ src/test/java/io/anserini/document/JsonDocumentArrayTest.java
@@ -23,11 +23,12 @@ import java.io.OutputStreamWriter;
 import java.io.Writer;
 import java.util.HashMap;
 
+import io.anserini.collection.JsonCollection;
 import org.junit.After;
 import org.junit.Before;
 
 
-public class JsonDocumentArrayTest extends DocumentTest<JsonDocument> {
+public class JsonDocumentArrayTest extends DocumentTest<JsonCollection.Document> {
   private String sampleFile = "sampleJsonArray.json";
 
   @Before
@@ -49,7 +50,7 @@ public class JsonDocumentArrayTest extends DocumentTest<JsonDocument> {
       new FileOutputStream(sampleFile), "utf-8"));
     writer.write(sampleDoc);
 
-    dType = new JsonDocument(sampleFile);
+    dType = new JsonCollection.Document(sampleFile);
 
     HashMap<String, String> doc1 = new HashMap<>();
     doc1.put("id", "doc1");
